after hours of debugging, i found the bug, next is ti fix the int

+++ Core/DRIVERS_V2.0/I2C/I2C_POLLING.h
@@ -13,7 +13,7 @@ typedef struct{
 void I2Cx_Polling_init(I2C_TypeDef *i2c, i2c__polling_config_t *config);
 void I2Cx_Polling_start(I2C_TypeDef *i2c);
 void I2Cx_Polling_write(I2C_TypeDef* i2c,uint8_t slave_addr, uint8_t register_addr, uint8_t *buffer, uint16_t length);
-void I2Cx_Polling_read(I2C_TypeDef *i2c,uint16_t slave_addr, uint16_t register_addr, uint8_t *buffer, uint8_t length);
+void I2Cx_Polling_read(I2C_TypeDef *i2c,uint8_t slave_addr, uint8_t register_addr, uint8_t *buffer, uint8_t length);
 void I2Cx_Polling_stop(I2C_TypeDef *i2c);
 
 #endif
